+++ src/main/java/koreatech/in/controller/OwnerController.java
@@ -13,8 +13,8 @@ import koreatech.in.annotation.AuthExcept;
 import koreatech.in.annotation.ParamValid;
 import koreatech.in.dto.EmptyResponse;
 import koreatech.in.dto.ExceptionResponse;
-import koreatech.in.dto.normal.owner.request.VerifyCodeRequest;
-import koreatech.in.dto.normal.owner.request.VerifyEmailRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyCodeRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyEmailRequest;
 import koreatech.in.exception.BaseException;
 import koreatech.in.exception.ExceptionInformation;
 import koreatech.in.service.OwnerService;
