+++ src/main/java/koreatech/in/mapstruct/OwnerConverter.java
@@ -4,8 +4,8 @@ import koreatech.in.domain.User.owner.Domain;
 import koreatech.in.domain.User.owner.EmailAddress;
 import koreatech.in.domain.User.owner.LocalParts;
 import koreatech.in.domain.User.owner.OwnerInCertification;
-import koreatech.in.dto.normal.owner.request.VerifyCodeRequest;
-import koreatech.in.dto.normal.owner.request.VerifyEmailRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyCodeRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyEmailRequest;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
