+++ src/main/java/koreatech/in/service/OwnerService.java
@@ -1,7 +1,7 @@
 package koreatech.in.service;
 
-import koreatech.in.dto.normal.owner.request.VerifyCodeRequest;
-import koreatech.in.dto.normal.owner.request.VerifyEmailRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyCodeRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyEmailRequest;
 
 public interface OwnerService {
