+++ src/main/java/koreatech/in/service/OwnerServiceImpl.java
@@ -10,8 +10,8 @@ import koreatech.in.domain.User.owner.CertificationCode;
 import koreatech.in.domain.User.owner.EmailAddress;
 import koreatech.in.domain.User.owner.OwnerInCertification;
 import koreatech.in.domain.User.owner.OwnerInVerification;
-import koreatech.in.dto.normal.owner.request.VerifyCodeRequest;
-import koreatech.in.dto.normal.owner.request.VerifyEmailRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyCodeRequest;
+import koreatech.in.dto.normal.user.owner.request.VerifyEmailRequest;
 import koreatech.in.exception.BaseException;
 import koreatech.in.exception.ExceptionInformation;
 import koreatech.in.mapstruct.OwnerConverter;
